+++ Sources/SwiftyHaru/Graphics/LineJoin.swift
@@ -13,7 +13,7 @@ import typealias CLibHaru.HPDF_LineJoin
 /// - miter: A join with a sharp (angled) corner.
 /// - round: A join with a rounded end.
 /// - bevel: A join with a squared-off end.
-public enum LineJoin: UInt32 {
+public enum LineJoin: UInt32, CaseIterable {
     
     /// A join with a sharp (angled) corner.
     ///
@@ -31,7 +31,7 @@ public enum LineJoin: UInt32 {
     case bevel
 }
 
-internal extension LineJoin {
+extension LineJoin {
     
     internal init(_ haruEnum: HPDF_LineJoin) {
         self.init(rawValue: haruEnum.rawValue)!