+++ Sources/SwiftyHaru/PDFDocument.PageNumberStyle.swift
@@ -8,10 +8,10 @@
 
 import CLibHaru
 
-public extension PDFDocument {
+extension PDFDocument {
     
     /// Determines a page labeling style of the document.
-    public enum PageNumberStyle: UInt32 {
+    public enum PageNumberStyle: UInt32, CaseIterable {
         
         /// Arabic numerals (1 2 3 4).
         case decimal = 0