+++ Sources/SwiftyHaru/PDFPage.Direction.swift
@@ -6,24 +6,15 @@
 //
 //
 
-import CLibHaru
-
-public extension PDFPage {
+extension PDFPage {
     
     /// Orientation of a page.
-    public enum Direction: UInt32 {
+    public enum Direction: UInt32, CaseIterable {
         
         /// The width is less than the height.
         case portrait
         
         /// The height is less than the width.
         case landscape
-    }
-}
-
-internal extension PDFPage.Direction {
-    
-    internal init?(haruEnum: HPDF_PageDirection) {
-        self.init(rawValue: haruEnum.rawValue)
     }
 }