+++ Sources/SwiftyHaru/TextAlignment.swift
@@ -9,7 +9,7 @@
 import CLibHaru
 
 /// The alignment of text to use in the `DrawingContext.show(text:in:alignment:)` method.
-public enum TextAlignment: UInt32 {
+public enum TextAlignment: UInt32, CaseIterable {
 
     /// The text is aligned to left.
     case left