+++ Tests/SwiftyHaruTests/DrawingContextTests.swift
@@ -42,17 +42,22 @@ final class DrawingContextTests: TestCase {
         ("testTextEncoding", testTextEncoding),
         ("testTextEncodingUnsupportedByCurrentFont", testTextEncodingUnsupportedByCurrentFont),
         ("testTextWidthForString", testTextWidthForString),
+        ("testMeasureText", testMeasureText),
         ("testTextBoundingBox", testTextBoundingBox),
         ("testFontAscent", testFontAscent),
         ("testFontDescent", testFontDescent),
         ("testFontXHeight", testFontXHeight),
         ("testFontCapHeight", testFontCapHeight),
         ("testTextLeading", testTextLeading),
+        ("testTextRenderingMode", testTextRenderingMode),
+        ("testCharacterSpacing", testCharacterSpacing),
+        ("testWordSpacing", testWordSpacing),
         // Text showing
         ("testShowOnelineText", testShowOnelineText),
         ("testShowMultilineText", testShowMultilineText),
         ("testShowUnicodeText", testShowUnicodeText),
-        ("testShowTextInRect", testShowTextInRect)
+        ("testShowTextInRect", testShowTextInRect),
+        ("testShowWithTextMatrix", testShowWithTextMatrix)
     ]
     
     // MARK: - Helpers
@@ -76,6 +81,26 @@ final class DrawingContextTests: TestCase {
         
         try grid.draw(in: context, position: .zero + gridOrigin)
         drawPoint(gridCenter, in: context)
+    }
+
+    private func drawStripePattern(in context: DrawingContext, rect: Rectangle) {
+        let savedLineWidth = context.lineWidth
+        let savedStrokeColor = context.strokeColor
+        defer {
+            context.lineWidth = savedLineWidth
+            context.strokeColor = savedStrokeColor
+        }
+
+        context.lineWidth = 1
+        context.strokeColor = #colorLiteral(red: 0.2196078449, green: 0.007843137719, blue: 0.8549019694, alpha: 1)
+
+        for dy in stride(from: 0 as Float, to: rect.maxY, by: 3) {
+            context.stroke(
+                Path()
+                    .moving(toX: rect.x, y: rect.y + dy)
+                    .appendingLine(toX: rect.maxX, y: rect.y + dy)
+            )
+        }
     }
     
     // MARK: - Graphics state
@@ -701,46 +726,54 @@ final class DrawingContextTests: TestCase {
     
     // MARK: - Text State
     
-    func testTextFont() {
+    func testTextFont() throws {
         
         // Given
         let expectedInitialFont = Font.helvetica
-        let expectedFont = Font.courierBold
 
         // When
         var returnedInitialFont: Font?
-        var returnedFont: Font?
 
-        document.addPage { context in
+        try document.addPage { context in
             returnedInitialFont = context.font
-            context.font = .courierBold
-            returnedFont = context.font
+
+            let ys = sequence(first: 100, next: { $0 + 1.2 * context.fontSize })
+            for (font, y) in zip(Font.baseFonts, ys) {
+                context.font = font
+                try context.show(text: font.name, atX: 100, y: y)
+            }
         }
 
         // Then
         XCTAssertEqual(expectedInitialFont, returnedInitialFont)
-        XCTAssertEqual(expectedFont, returnedFont)
+        assertPDFSnapshot()
+        assertPDFImageSnapshot(page: 1, named: "1")
     }
     
-    func testTextFontSize() {
+    func testTextFontSize() throws {
         
         // Given
         let expectedInitialFontSize: Float = 11
-        let expectedFontSize: Float = 64
 
         // When
         var returnedInitialFontSize: Float?
-        var returnedFontSize: Float?
 
-        document.addPage { context in
+        try document.addPage { context in
             returnedInitialFontSize = context.fontSize
-            context.fontSize = 64
-            returnedFontSize = context.fontSize
+
+            let fontSizes = stride(from: 1 as Float, to: 70, by: 5)
+            let ys = sequence(first: 100, next: { $0 + 1.2 * context.fontSize })
+
+            for (fontSize, y) in zip(fontSizes, ys) {
+                context.fontSize = fontSize
+                try context.show(text: "font size is \(Int(fontSize))", atX: 100, y: y)
+            }
         }
 
         // Then
         XCTAssertEqual(expectedInitialFontSize, returnedInitialFontSize)
-        XCTAssertEqual(expectedFontSize, returnedFontSize)
+        assertPDFSnapshot()
+        assertPDFImageSnapshot(page: 1, named: "1")
     }
     
     func testTextEncoding() {
@@ -800,6 +833,39 @@ final class DrawingContextTests: TestCase {
         // Then
         XCTAssertEqual(expectedWidth, returnedWidth)
         XCTAssertEqual(expectedWidthForMultilineText, returnedWidthForMultilineText)
+    }
+
+    func testMeasureText() throws {
+
+        // Given
+        let expectedLengthNoWordwrap = 12
+        let expectedWidthNoWordwrap: Float = 75.78999
+        let expectedLengthWordwrap = 10
+        let expectedWidthWordwrap: Float = 64.173996
+        let text = "ABCDE FGH IJKL"
+
+        // When
+        var returnedLengthNoWordwrap: Int?
+        var returnedWidthNoWordwrap: Float?
+        var returnedLengthWordwrap: Int?
+        var returnedWidthWordwrap: Float?
+
+        try document.addPage { context in
+
+            let noWordWrap = try context.measureText(text, width: 80, wordwrap: false)
+            returnedLengthNoWordwrap = noWordWrap.utf8Length
+            returnedWidthNoWordwrap  = noWordWrap.realWidth
+
+            let withWordWrap = try context.measureText(text, width: 80, wordwrap: true)
+            returnedLengthWordwrap = withWordWrap.utf8Length
+            returnedWidthWordwrap  = withWordWrap.realWidth
+        }
+
+        // Then
+        XCTAssertEqual(expectedLengthNoWordwrap, returnedLengthNoWordwrap)
+        XCTAssertEqual(expectedWidthNoWordwrap, returnedWidthNoWordwrap)
+        XCTAssertEqual(expectedLengthWordwrap, returnedLengthWordwrap)
+        XCTAssertEqual(expectedWidthWordwrap, returnedWidthWordwrap)
     }
     
     func testTextBoundingBox() {
@@ -939,6 +1005,110 @@ final class DrawingContextTests: TestCase {
         XCTAssertEqual(expectedInitialTextLeading, returnedInitialTextLeading)
         XCTAssertEqual(expectedTextLeading, returnedTextLeading)
     }
+
+    func testTextRenderingMode() throws {
+
+        // Given
+        let expectedInitialMode = TextRenderingMode.fill
+
+        // When
+        var returnedInitialMode: TextRenderingMode?
+
+        try document.addPage { context in
+
+            returnedInitialMode = context.textRenderingMode
+
+            context.fillColor = #colorLiteral(red: 0.9686274529, green: 0.78039217, blue: 0.3450980484, alpha: 1)
+            context.fontSize = 24
+
+            let ys = sequence(first: 100, next: { $0 + 2 * context.fontSize })
+            for (mode, y) in zip(TextRenderingMode.allCases, ys) {
+
+                try context.withNewGState {
+                    context.textRenderingMode = mode
+                    let text = String(describing: mode)
+                    try context.show(text: text, atX: 100, y: y)
+
+                    switch mode {
+                    case .fillClipping, .strokeClipping, .fillStrokeClipping, .clipping:
+                        drawStripePattern(in: context, rect: Rectangle(x: 100,
+                                                                       y: y,
+                                                                       width: context.textWidth(for: text),
+                                                                       height: context.fontCapHeight))
+                    default:
+                        break
+                    }
+                }
+            }
+        }
+
+        // Then
+        XCTAssertEqual(expectedInitialMode, returnedInitialMode)
+        assertPDFSnapshot()
+        assertPDFImageSnapshot(page: 1, named: "1")
+    }
+
+    func testCharacterSpacing() throws {
+
+        // Given
+        let expectedInitialSpacing: Float = 0
+
+        // When
+        var returnedInitialSpacing: Float?
+
+        try document.addPage { context in
+
+            returnedInitialSpacing = context.characterSpacing
+            context.fontSize = 7
+
+            let ys = sequence(first: 100, next: { $0 + 1.2 * context.fontSize })
+            let spacings = stride(from: -7 as Float,
+                                  to: 7,
+                                  by: 1)
+
+            for (spacing, y) in zip(spacings, ys) {
+                context.characterSpacing = spacing
+
+                try context.show(text: "All work and no play makes Jack a dull boy", atX: 170, y: y)
+            }
+        }
+
+        // Then
+        XCTAssertEqual(expectedInitialSpacing, returnedInitialSpacing)
+        assertPDFSnapshot()
+        assertPDFImageSnapshot(page: 1, named: "1")
+    }
+
+    func testWordSpacing() throws {
+
+        // Given
+        let expectedInitialSpacing: Float = 0
+
+        // When
+        var returnedInitialSpacing: Float?
+
+        try document.addPage { context in
+
+            returnedInitialSpacing = context.wordSpacing
+            context.fontSize = 7
+
+            let ys = sequence(first: 100, next: { $0 + 1.2 * context.fontSize })
+            let spacings = stride(from: DrawingContext.minWordSpacing,
+                                  to: 30,
+                                  by: 1)
+
+            for (spacing, y) in zip(spacings, ys) {
+                context.wordSpacing = spacing
+
+                try context.show(text: "All work and no play makes Jack a dull boy", atX: 160, y: y)
+            }
+        }
+
+        // Then
+        XCTAssertEqual(expectedInitialSpacing, returnedInitialSpacing)
+        assertPDFSnapshot()
+        assertPDFImageSnapshot(page: 1, named: "1")
+    }
     
     // MARK: - Text Showing
     
@@ -1047,5 +1217,29 @@ final class DrawingContextTests: TestCase {
 
         XCTAssertEqual(result2.charactersPrinted, text.count)
         XCTAssertTrue(result2.isSufficientSpace)
+    }
+
+    func testShowWithTextMatrix() throws {
+
+        // Given
+        let matrices = [
+            AffineTransform.identity,
+            AffineTransform(translationX: 210, y: 160).scaled(byX: 1, y: 2),
+            AffineTransform(translationX: 210, y: 220).rotated(byAngle: .pi / 6),
+            AffineTransform(a: 1, b: tan(.pi / 18), c: tan(.pi / 9), d: 1, tx: 210, ty: 280),
+        ]
+
+        // When
+        try document.addPage { context in
+
+            let ys = sequence(first: 100 as Float, next: { $0 + 60 })
+            for (matrix, y) in zip(matrices, ys) {
+                try context.show(text: "Hello World!", atX: 200, y: y, textMatrix: matrix)
+            }
+        }
+
+        // Then
+        assertPDFSnapshot()
+        assertPDFImageSnapshot(page: 1, named: "1")
     }
 }